+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- MSSQL & Azure MSSQL: The connect timeout when retrieving the connector form now defaults to 5 seconds
+
 ## [8.0.1] 2025-04-14
 
 ### Fixed